+++ Sources/StreamChatSwiftUI/ChatMessageList/DefaultMessageAttachmentsViewModifier.swift
@@ -42,8 +42,8 @@ struct DefaultMessageAttachmentsViewModifier<Style: Styles>: ViewModifier {
 
     private var bubbleInsets: EdgeInsets {
         guard isBubbleShown else { return EdgeInsets() }
-        // Single voice and file don't have extra padding.
-        if options.message.hasSingleAttachment(of: [.file, .voiceRecording], captioned: false) {
+        // Single voice, file, and audio don't have extra padding.
+        if options.message.hasSingleAttachment(of: [.file, .audio, .voiceRecording], captioned: false) {
             return EdgeInsets()
         }
         return EdgeInsets(
@@ -71,9 +71,9 @@ struct DefaultMessageAttachmentItemViewModifier: ViewModifier {
     @ViewBuilder
     func body(content: Content) -> some View {
         switch options.attachmentType {
-        case .some(.file):
+        case .some(.file), .some(.audio):
             content.modifier(BubbleModifier(
-                corners: attachmentCorners(isSingleWithoutCaption: options.message.hasSingleAttachment(of: [.file], captioned: false)),
+                corners: attachmentCorners(isSingleWithoutCaption: options.message.hasSingleAttachment(of: [.file, .audio], captioned: false)),
                 backgroundColors: [defaultAttachmentBackgroundColor],
                 borderWidth: 0,
                 cornerRadius: tokens.messageBubbleRadiusAttachment,
@@ -125,14 +125,14 @@ struct DefaultMessageAttachmentItemViewModifier: ViewModifier {
                 contentInsets: EdgeInsets()
             ))
         default:
-            // Other attachment types (e.g. giphy, audio, custom) are not wrapped in a container bubble.
+            // Other attachment types (e.g. giphy, custom) are not wrapped in a container bubble.
             content
         }
     }
 
     private var defaultAttachmentBackgroundColor: Color {
-        // Single file and voice attachments are rendered in a bubble, but the attachment itself does not have an additional darker background.
-        if options.message.hasSingleAttachment(of: [.file, .voiceRecording], captioned: false) {
+        // Single file, audio, and voice attachments are rendered in a bubble, but the attachment itself does not have an additional darker background.
+        if options.message.hasSingleAttachment(of: [.file, .audio, .voiceRecording], captioned: false) {
             return .clear
         }
         return Color(options.message.isSentByCurrentUser ? colors.chatBackgroundAttachmentOutgoing : colors.chatBackgroundAttachmentIncoming)